refactor: move LostCalleeSavedRegisterCheck to own file
